Fix race condition in Utils + remove duplicated code
