+++ src/tasks/AotCompilerTask/MonoAOTCompiler.csproj
@@ -15,7 +15,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="MonoAOTCompiler.cs" />
-    <Compile Include="Utils.cs" />
+    <Compile Include="..\Common\Utils.cs" />
   </ItemGroup>
   <ItemGroup>
     <Content Include="MonoAOTCompiler.props">
