+++ src/tasks/AppleAppBuilder/TargetOS.cs
@@ -1,0 +1,8 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+internal class TargetOS
+{
+    public const string iOS = "iOS";
+    public const string tvOS = "tvOS";
+}
